move cad model helper under get platform config
