+++ lib/getPlatformConfig.ts
@@ -6,8 +6,8 @@ import {
 } from "@tscircuit/parts-engine"
 import { createKiCadRoutingToolsAutorouter } from "@tscircuit/krt-wasm"
 import { parseKicadModToCircuitJson } from "kicad-component-converter"
+import { extractCadModelFromCircuitJson } from "./getPlatformConfig/extractCadModelFromCircuitJson"
 import { dynamicallyLoadDependencyWithCdnBackup } from "./utils/dynamically-load-dependency-with-cdn-backup"
-import { extractCadModelFromCircuitJson } from "./utils/extractCadModelFromCircuitJson"
 import { KicadToCircuitJsonConverter } from "kicad-to-circuit-json"
 import type { AnyCircuitElement } from "circuit-json"
 import * as React from "react"
