+++ lib/getPlatformConfig/extractCadModelFromCircuitJson.ts
